Fix indexing of staked balances.
This updates the staked balances indexer so at all times the
sacredness of unique IDs is preserved.

Two schemas are used for this. In one, I store a rolling balance per
address, per contract, and in the other I store a rolling balance per
address, per contract, per block height.

You need a rolling balance, per contract, per address, per block
height because it is possible two stake transactions for an address
could occur in the same block. If a rolling balance was not stored
only the second update would be reflected.